refactor: item, coin 도메인 분리
